// frontend/src/app/core/api/models/lowCostFlightAuthCommonAuthResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SkyDealRadar API
 * OpenAPI spec version: v1
 */
import type { LowCostFlightAuthCommonUserDto } from './lowCostFlightAuthCommonUserDto';

export interface LowCostFlightAuthCommonAuthResponse {
  token?: string;
  expiresAt?: string;
  user?: LowCostFlightAuthCommonUserDto;
}